+++ mux/modules/engine/jit_compiler.cpp
@@ -4533,7 +4533,7 @@ FUNCTION(fun__restore_cargs)
 
     if (!s_current_ecall_ctx || nfargs < 1) return;
 
-    int idx = mux_atoi64(fargs[0]);
+    int64_t idx = mux_atoi64(fargs[0]);
     if (idx >= 0 && idx < MAX_CARG_SAVE_DEPTH
         && s_carg_save_stack[idx].in_use)
     {
@@ -4576,7 +4576,7 @@ FUNCTION(fun__set_ncargs)
 
     if (!s_current_ecall_ctx || nfargs < 1) return;
 
-    int n = mux_atoi64(fargs[0]);
+    int64_t n = mux_atoi64(fargs[0]);
     if (n < 0) n = 0;
     if (n > 10) n = 10;
 
@@ -4607,7 +4607,7 @@ FUNCTION(fun__write_carg)
 
     if (!s_current_ecall_ctx || nfargs < 2) return;
 
-    int idx = mux_atoi64(fargs[0]);
+    int64_t idx = mux_atoi64(fargs[0]);
     if (idx < 0 || idx >= 10) return;
 
     uint64_t dst = rv_compiler::CARGS_BASE
