+++ mux/modules/mail/mail_mod.h
@@ -171,11 +171,13 @@ typedef unsigned int mail_flag;
 
 struct mail_selector
 {
-    int       low;
-    int       high;
+    // Message index / age filters: full-width integers (#1402).
+    //
+    int64_t   low;
+    int64_t   high;
     mail_flag flags;
     dbref     player;
-    int       days;
+    int64_t   days;
     int       day_comp;
 };
